+++ src/main/java/org/liubov/statetaxcalculator/service/impl/TaxCalculatorServiceImpl.java
@@ -25,7 +25,7 @@ public class TaxCalculatorServiceImpl implements TaxCalculatorService {
     }
 
     @Override
-    public double calculateStateTax(int year, String state, String filingStatus, int income) throws IncomeTaxCalculatorException {
+    public double calculateStateTax(int year, String state, String filingStatus, double income) throws IncomeTaxCalculatorException {
         if (state.equals("Alaska") || state.equals("Florida") || state.equals("Nevada")
                 || state.equals("New Hampshire") || state.equals("South Dakota") || state.equals("Tennessee")
                 || state.equals("Texas") || state.equals("Washington") || state.equals("Wyoming")) {
@@ -50,7 +50,7 @@ public class TaxCalculatorServiceImpl implements TaxCalculatorService {
         if (stateTaxBracketList != null && stateTaxBracketList.size() > 0) {
             for (StateTaxBracket stateTaxBracket : stateTaxBracketList) {
                 if (income >= stateTaxBracket.getBracketLower()) {
-                    return Math.round((stateTaxBracket.getAccumulatedAmount() + (income - stateTaxBracket.getBracketLower()) * stateTaxBracket.getRate() / 100) * 100) / 100;
+                    return Math.round((stateTaxBracket.getAccumulatedAmount() + (income - stateTaxBracket.getBracketLower()) * stateTaxBracket.getRate() / 100) * 100) / 100.0;
                 }
             }
         }
@@ -59,7 +59,7 @@ public class TaxCalculatorServiceImpl implements TaxCalculatorService {
     }
 
     @Override
-    public double calculateFederalTax(int year, String filingStatus, int income) throws IncomeTaxCalculatorException {
+    public double calculateFederalTax(int year, String filingStatus, double income) throws IncomeTaxCalculatorException {
         if (year != 2022 || year != 2023) {
             year = 2023;
         }
@@ -69,7 +69,7 @@ public class TaxCalculatorServiceImpl implements TaxCalculatorService {
         if (federalTaxBracketList != null && federalTaxBracketList.size() > 0) {
             for (FederalTaxBracket federalTaxBracket : federalTaxBracketList) {
                 if (income >= federalTaxBracket.getBracketLower()) {
-                    return Math.round((federalTaxBracket.getAccumulatedAmount() + (income - federalTaxBracket.getBracketLower()) * federalTaxBracket.getRate() / 100) * 100) / 100;
+                    return Math.round((federalTaxBracket.getAccumulatedAmount() + (income - federalTaxBracket.getBracketLower()) * federalTaxBracket.getRate() / 100) * 100) / 100.0;
                 }
             }
         }
@@ -78,31 +78,31 @@ public class TaxCalculatorServiceImpl implements TaxCalculatorService {
     }
 
     @Override
-    public double calculateFicaTax(int year, int income) {
+    public double calculateFicaTax(int year, double income) {
         if (year == 2023) {
             if (income < 160_200) {
-                return Math.round(income * 7.65) / 100;
+                return Math.round(income * 7.65) / 100.0;
             } else {
-                return Math.round(160_200 * 7.65) / 100;
+                return Math.round(160_200 * 7.65) / 100.0;
             }
         } else {
             if (income < 147_000) {
-                return Math.round(income * 7.65) / 100;
+                return Math.round(income * 7.65) / 100.0;
             } else {
-                return Math.round(147_000 * 7.65) / 100;
+                return Math.round(147_000 * 7.65) / 100.0;
             }
         }
     }
 
     @Override
-    public double calculateAdditionalMedicareTax(int year, String filingStatus, int income) {
+    public double calculateAdditionalMedicareTax(int year, String filingStatus, double income) {
         if (filingStatus.equals("Single") || filingStatus.equals("Married filing separately")) {
             if (income > 200_000) {
-                return Math.round((income - 200_000) * 0.9) / 100;
+                return Math.round((income - 200_000) * 0.9) / 100.0;
             }
         } else {
             if (income > 250_000) {
-                return Math.round((income - 250_000) * 0.9) / 100;
+                return Math.round((income - 250_000) * 0.9) / 100.0;
             }
         }
 
@@ -110,7 +110,7 @@ public class TaxCalculatorServiceImpl implements TaxCalculatorService {
     }
 
     @Override
-    public double calculateEffectiveTaxRate(double tax, int income) {
-        return Math.round(tax / income * 100 * 100) / 100;
+    public double calculateEffectiveTaxRate(double tax, double income) {
+        return Math.round(tax / income * 100 * 100) / 100.0;
     }
 }
